uint => uint256

+++ src/ZivoeDAO.sol
@@ -140,28 +140,28 @@ contract ZivoeDAO is ERC1155Holder, ERC721Holder, Ownable {
     /// @param  asset The asset being pushed.
     /// @param  tokenId The ID for a given "asset" / NFT.
     /// @param  data Accompanying data for the transaction.
-    event PushedERC721(address indexed locker, address indexed asset, uint indexed tokenId, bytes data);
+    event PushedERC721(address indexed locker, address indexed asset, uint256 indexed tokenId, bytes data);
 
     /// @notice Emitted during pushMultiERC721().
     /// @param  locker The locker receiving "assets".
     /// @param  assets The assets being pushed, corresponds to "tokenIds".
     /// @param  tokenIds The ID for a given "asset" / NFT.
     /// @param  data Accompanying data for the transaction(s).
-    event PushedMultiERC721(address indexed locker, address[] assets, uint[] tokenIds, bytes[] data);
+    event PushedMultiERC721(address indexed locker, address[] assets, uint256[] tokenIds, bytes[] data);
     
     /// @notice Emitted during pullERC721().
     /// @param  locker The locker "assets" are pulled from.
     /// @param  asset The asset being pulled.
     /// @param  tokenId The ID for a given "asset" / NFT.
     /// @param  data Accompanying data for the transaction.
-    event PulledERC721(address indexed locker, address indexed asset, uint indexed tokenId, bytes data);
+    event PulledERC721(address indexed locker, address indexed asset, uint256 indexed tokenId, bytes data);
 
     /// @notice Emitted during pullMultiERC721().
     /// @param  locker The locker "assets" are pulled from.
     /// @param  assets The assets being pulled.
     /// @param  tokenIds The ID for a given "asset" / NFT.
     /// @param  data Accompanying data for the transaction(s).
-    event PulledMultiERC721(address indexed locker, address[] assets, uint[] tokenIds, bytes[] data);
+    event PulledMultiERC721(address indexed locker, address[] assets, uint256[] tokenIds, bytes[] data);
     
     /// @notice Emitted during pushERC1155Batch().
     /// @param  locker The locker receiving "assets".
@@ -232,11 +232,11 @@ contract ZivoeDAO is ERC1155Holder, ERC721Holder, Ownable {
         require(assets.length == amounts.length, "ZivoeDAO::pushMulti() assets.length != amounts.length");
         require(IERC104_P_0(locker).canPushMulti(), "ZivoeDAO::pushMulti() !IERC104_P_0(locker).canPushMulti()");
         emit PushedMulti(locker, assets, amounts);
-        for (uint i = 0; i < assets.length; i++) {
+        for (uint256 i = 0; i < assets.length; i++) {
             IERC20(assets[i]).safeApprove(locker, amounts[i]);
         }
         IERC104_P_0(locker).pushToLockerMulti(assets, amounts);
-        for (uint i = 0; i < assets.length; i++) {
+        for (uint256 i = 0; i < assets.length; i++) {
             if (IERC20(assets[i]).allowance(address(this), locker) > 0) {
                 IERC20(assets[i]).safeApprove(locker, 0);
             }
@@ -268,7 +268,7 @@ contract ZivoeDAO is ERC1155Holder, ERC721Holder, Ownable {
     /// @param  asset The NFT contract.
     /// @param  tokenId The NFT ID to push.
     /// @param  data Accompanying data for the transaction.
-    function pushERC721(address locker, address asset, uint tokenId, bytes calldata data) external onlyOwner {
+    function pushERC721(address locker, address asset, uint256 tokenId, bytes calldata data) external onlyOwner {
         require(IZivoeGlobals_P_5(GBL).isLocker(locker), "ZivoeDAO::pushERC721() !IZivoeGlobals_P_5(GBL).isLocker(locker)");
         require(IERC104_P_0(locker).canPushERC721(), "ZivoeDAO::pushERC721() !IERC104_P_0(locker).canPushERC721()");
         emit PushedERC721(locker, asset, tokenId, data);
@@ -281,13 +281,13 @@ contract ZivoeDAO is ERC1155Holder, ERC721Holder, Ownable {
     /// @param  assets The NFT contracts.
     /// @param  tokenIds The NFT IDs to push.
     /// @param  data Accompanying data for the transaction(s).
-    function pushMultiERC721(address locker, address[] calldata assets, uint[] calldata tokenIds, bytes[] calldata data) external onlyOwner {
+    function pushMultiERC721(address locker, address[] calldata assets, uint256[] calldata tokenIds, bytes[] calldata data) external onlyOwner {
         require(IZivoeGlobals_P_5(GBL).isLocker(locker), "ZivoeDAO::pushERC721() !IZivoeGlobals_P_5(GBL).isLocker(locker)");
         require(assets.length == tokenIds.length, "ZivoeDAO::pushMultiERC721() assets.length != tokenIds.length");
         require(tokenIds.length == data.length, "ZivoeDAO::pushMultiERC721() tokenIds.length != data.length");
         require(IERC104_P_0(locker).canPushMultiERC721(), "ZivoeDAO::pushMultiERC721() !IERC104_P_0(locker).canPushMultiERC721()");
         emit PushedMultiERC721(locker, assets, tokenIds, data);
-        for (uint i = 0; i < assets.length; i++) {
+        for (uint256 i = 0; i < assets.length; i++) {
             IERC721_P_0(assets[i]).approve(locker, tokenIds[i]);
         }
         IERC104_P_0(locker).pushToLockerMultiERC721(assets, tokenIds, data);
@@ -298,7 +298,7 @@ contract ZivoeDAO is ERC1155Holder, ERC721Holder, Ownable {
     /// @param  asset The NFT contract.
     /// @param  tokenId The NFT ID to pull.
     /// @param  data Accompanying data for the transaction.
-    function pullERC721(address locker, address asset, uint tokenId, bytes calldata data) external onlyOwner {
+    function pullERC721(address locker, address asset, uint256 tokenId, bytes calldata data) external onlyOwner {
         require(IERC104_P_0(locker).canPullERC721(), "ZivoeDAO::pullERC721() !IERC104_P_0(locker).canPullERC721()");
         emit PulledERC721(locker, asset, tokenId, data);
         IERC104_P_0(locker).pullFromLockerERC721(asset, tokenId, data);
@@ -309,7 +309,7 @@ contract ZivoeDAO is ERC1155Holder, ERC721Holder, Ownable {
     /// @param  assets The NFT contracts.
     /// @param  tokenIds The NFT IDs to pull.
     /// @param  data Accompanying data for the transaction(s).
-    function pullMultiERC721(address locker, address[] calldata assets, uint[] calldata tokenIds, bytes[] calldata data) external onlyOwner {
+    function pullMultiERC721(address locker, address[] calldata assets, uint256[] calldata tokenIds, bytes[] calldata data) external onlyOwner {
         require(IERC104_P_0(locker).canPullMultiERC721(), "ZivoeDAO::pullMultiERC721() !IERC104_P_0(locker).canPullMultiERC721()");
         require(assets.length == tokenIds.length, "ZivoeDAO::pullMultiERC721() assets.length != tokenIds.length");
         require(tokenIds.length == data.length, "ZivoeDAO::pullMultiERC721() tokenIds.length != data.length");
